Fix LMS/distributed-training timing issue
During distributed training, it's possible that one thread can be destroying
a tensor while another thread is attempting to reclaim that same tensor's storage
in order to satisfy a new allocation.

This scenario is not a problem in itself, however in the current code the
reclaiming thread can encounter a runtime error if the destroying thread has
removed the block from the allocated_blocks map while the storage still resides
on the reclaim_list.

The fix is to simply ensure that the LMS state is torn down (and the storage removed
from the reclaim list) prior to releasing the memory back to the allocator.